Object development

+++ src/VertexArray.cpp
@@ -32,7 +32,7 @@ void VertexArray::AddBuffer(const class VertexBuffer &vb, const VertexBufferLayo
     for(unsigned int i = 0; i < elements.size(); i++ ){
         const auto &element = elements[i]
         GLCall(glEnableVertexAttribArray(i));
-        GLCall(glVertexAttribPointer(i,element.count, element.type, element.normalized, layout.GetStride() , (const void*)offset ));
+        GLCall(glVertexAttribPointer(i,element.count, element.type, element.normalized, layout.GetStride() , (const void*)(size_t)offset ));
         offset += element.count *  VertexBufferElements::GetSizeOfType(element.type);
     }
